Moved quillView.tsx and removed excess files
