+++ examples/data-objects/text-editor/src/index.ts
@@ -7,9 +7,9 @@ import { ContainerViewRuntimeFactory } from "@fluid-example/example-utils";
 import type { IReactTreeDataObject } from "@fluidframework/react/alpha";
 import * as React from "react";
 
+import { MainView } from "./quillView.js";
 import type { TextAsTree } from "./schema.js";
 import { TextEditorFactory } from "./textEditorFactory.js";
-import { MainView } from "./view/index.js";
 
 export const fluidExport = new ContainerViewRuntimeFactory(
 	TextEditorFactory,
