--- examples/data-objects/text-editor/src/view/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { MainView } from "./quill/index.js";
-export type { MainViewProps } from "./quill/index.js";
